requirements.txt
apache-airflow
fairlearn
scikit-learn
pandas
numpy
google-cloud-storage
google-auth
matplotlib
seaborn
hyperopt
mlflow
pytest
pandera
xgboost
SQLAlchemy
uvicorn


